pyproject.toml
[tool.poetry]
name = "discord-message-summarizer"
version = "0.1.0"
description = "A Discord bot that summarizes messages using Anthropic's Claude API"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
discord = "^2.0.0"
anthropic = "^0.34.1"
python-dotenv = "^0.19.0"

[tool.poetry.dev-dependencies]
pytest = "^7.0"

[tool.poetry.scripts]
summarize = "src.discord_message_summarizer:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
